Move updateConfigState from AddSession to utils/stateReducers for reuse
